@@ -5,19 +5,17 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.2.2.RELEASE</version>
+        <version>3.2.0</version>
         <relativePath/>
     </parent>
-    <groupId>com</groupId>
+    <groupId>userService</groupId>
     <artifactId>User</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>User</name>
     <description>Demo project for Spring Boot</description>
 
     <properties>
-        <java.version>1.8</java.version>
-        <spring-cloud.version>2020.0.3</spring-cloud.version>
-        <hibernate.version>5.4.32.Final</hibernate.version>
+        <java.version>18</java.version>
         <swagger.version>2.9.2</swagger.version>
     </properties>
 
