@@ -1,6 +1,7 @@
 package com.User.domain;
 
-import javax.persistence.*;
+
+import jakarta.persistence.*;
 
 @Entity
 @Table(name = "Role")
