@@ -1,13 +1,16 @@
 package raf.mapper;
 
+import com.User.repository.ManagerRepository;
+import lombok.AllArgsConstructor;
 import raf.domain.FiskulturnaSala;
 import raf.dto.FiskulturnaSalaCreateDto;
 import raf.dto.FiskulturnaSalaDto;
 import org.springframework.stereotype.Component;
 
 @Component
+@AllArgsConstructor
 public class FiskulturnaSalaMapper {
-
+    private ManagerRepository managerRepository;
     public FiskulturnaSalaDto DomainObjectToDto(FiskulturnaSala fs){
         FiskulturnaSalaDto fsDto = new FiskulturnaSalaDto();
         fsDto.setId(fs.getSala_id());
@@ -20,11 +23,13 @@ public class FiskulturnaSalaMapper {
     }
 
     public FiskulturnaSala DtoToDomainObject(FiskulturnaSalaCreateDto fsDto){
+
         FiskulturnaSala fs = new FiskulturnaSala();
         fs.setName(fsDto.getNaziv());
         fs.setOpis(fsDto.getOpis());
         fs.setKapacitet(fsDto.getKapacitet());
         fs.setBrojTrenera(fsDto.getBrojTrenera());
+        fs.setManager(managerRepository.findById(fsDto.getManager_id()).orElseThrow(() ->new RuntimeException()));
 
         return fs;
     }
