@@ -6,16 +6,13 @@ import com.User.dto.TokenRequestDto;
 import com.User.dto.TokenResponseDto;
 import com.User.security.CheckSecurity;
 import com.User.service.AdminService;
-import io.swagger.annotations.ApiImplicitParam;
-import io.swagger.annotations.ApiImplicitParams;
-import io.swagger.annotations.ApiOperation;
+import jakarta.validation.Valid;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
 
 @RestController
 @RequestMapping("/admin")
@@ -26,7 +23,7 @@ public class AdminController {
         this.adminService = adminService;
     }
 
-    @ApiOperation(value = "Ban client")
+
     @PostMapping("/ban/{userId}")
     @CheckSecurity(roles={"ROLE_ADMIN"})
     public ResponseEntity<String> banUser(@RequestHeader("Authorization") String authorization, @RequestBody @Valid @PathVariable Long userId) {
@@ -35,7 +32,7 @@ public class AdminController {
         return ResponseEntity.ok("Client with ID " + userId + " has been banned.");
     }
 
-    @ApiOperation(value = "Unban client")
+
     @PostMapping("/unban/{userId}")
     @CheckSecurity(roles={"ROLE_ADMIN"})
     public ResponseEntity<String> unbanUser(@RequestHeader("Authorization") String authorization,@RequestBody @Valid @PathVariable Long userId) {
@@ -44,14 +41,7 @@ public class AdminController {
         return ResponseEntity.ok("Client with ID " + userId + " has been unbanned.");
     }
 
-    @ApiOperation(value = "Get all admins")
-    @ApiImplicitParams({
-            @ApiImplicitParam(name = "page", value = "What page number you want", dataType = "string", paramType = "query"),
-            @ApiImplicitParam(name = "size", value = "Number of items to return", dataType = "string", paramType = "query"),
-            @ApiImplicitParam(name = "sort", allowMultiple = true, dataType = "string", paramType = "query",
-                    value = "Sorting criteria in the format: property(,asc|desc). " +
-                            "Default sort order is ascending. " +
-                            "Multiple sort criteria are supported.")})
+
     @GetMapping
     @CheckSecurity(roles = {"ROLE_ADMIN"})
     public ResponseEntity<Page<AdminDto>> getAllAdmins(@RequestHeader("Authorization") String authorization,
@@ -60,7 +50,6 @@ public class AdminController {
         return new ResponseEntity<>(adminService.findAll(pageable), HttpStatus.OK);
     }
 
-    @ApiOperation(value = "Login")
     @PostMapping("/login")
     public ResponseEntity<TokenResponseDto> loginUser(@RequestBody @Valid TokenRequestDto tokenRequestDto) {
         return new ResponseEntity<>(adminService.login(tokenRequestDto), HttpStatus.OK);
