@@ -7,16 +7,13 @@ import com.User.dto.TokenRequestDto;
 import com.User.dto.TokenResponseDto;
 import com.User.security.CheckSecurity;
 import com.User.service.ManagerService;
-import io.swagger.annotations.ApiImplicitParam;
-import io.swagger.annotations.ApiImplicitParams;
-import io.swagger.annotations.ApiOperation;
+import jakarta.validation.Valid;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
 @RestController
 @RequestMapping("/manager")
 public class ManagerController {
@@ -26,14 +23,6 @@ public class ManagerController {
         this.managerService = managerService;
     }
 
-    @ApiOperation(value = "Get all managers")
-    @ApiImplicitParams({
-            @ApiImplicitParam(name = "page", value = "What page number you want", dataType = "string", paramType = "query"),
-            @ApiImplicitParam(name = "size", value = "Number of items to return", dataType = "string", paramType = "query"),
-            @ApiImplicitParam(name = "sort", allowMultiple = true, dataType = "string", paramType = "query",
-                    value = "Sorting criteria in the format: property(,asc|desc). " +
-                            "Default sort order is ascending. " +
-                            "Multiple sort criteria are supported.")})
     @GetMapping
    @CheckSecurity(roles = {"ROLE_MANAGER"})
     public ResponseEntity<Page<ManagerDto>> getAllMangers(@RequestHeader("Authorization") String authorization,
@@ -42,14 +31,14 @@ public class ManagerController {
         return new ResponseEntity<>(managerService.findAll(pageable), HttpStatus.OK);
     }
 
-    @ApiOperation(value = "Register manager")
+
     @PostMapping
     public ResponseEntity<ManagerDto> saveManager(@RequestBody @Valid ManagerCreateDto managerCreateDto) {
         //notifikacija
         return new ResponseEntity<>(managerService.add(managerCreateDto), HttpStatus.CREATED);
     }
 
-    @ApiOperation(value = "Login")
+
     @PostMapping("/login")
     public ResponseEntity<TokenResponseDto> loginManager(@RequestBody @Valid TokenRequestDto tokenRequestDto) {
         //notifikacija
