@@ -1,7 +1,9 @@
 package com.User.domain;
 
-import javax.persistence.*;
+//import javax.persistence.*;
 
+
+import jakarta.persistence.*;
 
 @Entity
 @Table(name = "Client")
